Implement ⟨ab⟩ syntax for taking output mid-predicate
The semantics here are that ⟨ab⟩ is equivalent to {[I,J]∧Ia.bJ∧},
i.e. it changes a predicate to take both input and output from the
left, and specify a "midpoint" for producing the connection to the
right.

This operation needs a better name, but it seems highly useful for
situation where we have two inputs and one output.